mail and permission in backend

+++ backend/src/resolvers/Mutation.js
@@ -2,6 +2,8 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
+const { makeANiceEmail, sgMail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const dayExpire = 1000 * 60 * 60 * 24;
 
@@ -15,9 +17,17 @@ signin = (userId, ctx) => {
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
     const item = await ctx.db.mutation.createItem(
       {
         data: {
+          user: {
+            connect: {
+              id: ctx.request.userId
+            }
+          },
           ...args
         }
       },
@@ -40,8 +50,15 @@ const Mutations = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    const item = await ctx.db.query.item({ where }, `{ id title}`);
-    console.log(item);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMINS', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that");
+    }
     return ctx.db.mutation.deleteItem({ where }, info);
   },
   async signup(parent, args, ctx, info) {
@@ -93,8 +110,19 @@ const Mutations = {
       where: { email: args.email },
       data: { resetToken, resetTokenExpiry }
     });
-    console.log(response);
-    return { message: 'Sign out success' };
+    const urlReset = `<a href="${
+      process.env.FRONTEND_URL
+    }/reset?resetToken=${resetToken}">Click here to reset</a>`;
+    const mailResponse = await sgMail.send({
+      from: 'dev@mloidi.com',
+      to: user.email,
+      subject: 'Your password reset link',
+      html: makeANiceEmail(
+        `Your password reset token is here! \n\n ${urlReset}`
+      )
+    });
+
+    return { message: 'Email send' };
   },
   async resetPassword(parent, args, ctx, info) {
     if (args.password !== args.confirmPassword) {
@@ -121,6 +149,31 @@ const Mutations = {
     });
     this.signin(updatedUser.id, ctx);
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in!');
+    }
+    const currentUser = await ctx.db.query.user(
+      {
+        where: { id: ctx.request.userId }
+      },
+      info
+    );
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
